DisplayIslandState optimization function via stream
Added several animal classes for testing

Changes have been made to the pet food function

+++ src/main/java/javaRush/Herbivore.java
@@ -3,7 +3,7 @@ package javaRush;
 import java.util.List;
 
 public abstract class Herbivore extends Animal {
-    public Herbivore(int weight, int maxCountPerLocation, int moveSpeed, int foodNeeded) {
+    public Herbivore(double weight, int maxCountPerLocation, int moveSpeed, double foodNeeded) {
         super(weight, maxCountPerLocation, moveSpeed, foodNeeded);
     }
